requirements.txt
flask==3.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.26.3
matplotlib==3.8.2
scikit-learn==1.3.2
openai==1.6.1
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
langchain==0.1.6
python-dateutil==2.8.2
websockets==10.3